Simplify mobile nav
