--- src/components/MenuItem.js
@@ -1,46 +1,0 @@
-import React from "react"
-import { Link } from "gatsby"
-import { motion } from "framer-motion"
-import styled from "styled-components"
-
-const variants = {
-  open: {
-    y: 0,
-    opacity: 1,
-    transition: {
-      y: { stiffness: 1000, velocity: -100 },
-    },
-  },
-  closed: {
-    y: 50,
-    opacity: 0,
-    transition: {
-      y: { stiffness: 1000 },
-    },
-  },
-}
-
-const Item = styled(motion.li)`
-  list-style: none;
-  margin-bottom: 20px;
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-`
-
-const MenuItem = ({ item, toggle }) => {
-  return (
-    <Item
-      variants={variants}
-      whileHover={{ scale: 1.1 }}
-      whileTap={{ scale: 0.95 }}
-      onClick={toggle}
-    >
-      <h3>
-        <Link to={item.route}>{item.text}</Link>
-      </h3>
-    </Item>
-  )
-}
-
-export default MenuItem
